+++ deployer/db/task/db:export.php
@@ -80,4 +80,4 @@ task('db:export', function () {
     if ($returnDumpCode) {
         writeln(json_encode(['dumpCode' => $dumpCode]));
     }
-})->desc('Export database dump to local database dumps storage.');
+})->desc('Export database dumps to current database dumps storage.');