+++ deployer/db/task/db:import.php
@@ -137,4 +137,4 @@ task('db:import', function () {
         $activePath = get('deploy_path') . '/' . (test('[ -L {{deploy_path}}/release ]') ? 'release' : 'current');
         run('cd ' . $activePath . ' && {{bin/php}} {{bin/deployer}} db:import --dumpcode=' . $dumpCode . ' ' . $verbosity);
     }
-})->desc('Import the database with "dumpcode" from current database dumps storage to database.');
+})->desc('Import dumps with given dumpcode from current database dumps storage to database.');