+++ deployer/db/task/db:upload.php
@@ -24,4 +24,4 @@ task('db:upload', function () {
         escapeshellarg($fileUtility->normalizeFolder(get('db_current_server')->get('db_storage_path_current'))),
         escapeshellarg($rsyncUtility->getHostWithDbStoragePath(Task\Context::get()))
     ), 0);
-})->desc('Upload the latest database dump from target database dumps storage.');
+})->desc('Upload the database dumps for given dumpcode from current to target database dumps storage.');