+++ deployer/db/config/set.php
@@ -5,12 +5,12 @@ namespace Deployer;
 use SourceBroker\DeployerExtendedDatabase\Utility\ArrayUtility;
 use SourceBroker\DeployerExtendedDatabase\Utility\InstanceUtility;
 
-// deployer settings
+// Deployer standard settings. By setting 'default_stage' you can do 'dep db:backup' instead of 'dep db:backup local'
 set('default_stage', function () {
     return (new InstanceUtility)->getCurrentInstance();
 });
 
-// Return current instance name. Based on that scripts knows from which server() takes the data to database.
+// Return current instance name. Based on that scripts knows from which server() takes the data to database operations.
 set('db_instance', function () {
     return (new InstanceUtility)->getCurrentInstance();
 });