+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 4) [TASK] Improve tasks descriptions.
 5) [FEATURE] Add db:compress and db:decompress tasks and extend docs.
 6) [TASK] Cleanup for db:upload, db:download tasks.
+7) [FEATURE] Compress local dumps after importing them with 'db:pull [instance]'.
 
 
 5.0.4