+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 7) [FEATURE] Compress local dumps after importing them with 'db:pull [instance]'.
 8) [FEATURE] Add db:rmdump task and documentation.
 9) [FEATURE] Add db:rmdump task at the end of "db:copy [source] [target]" task.
+10) [FEATURE] Add db:backup task.
 
 
 5.0.4